pyproject.toml
[project]
name = "chord"
version = "0.1.0"
description = "Chord: Automatic XTHP Threat Analyzer — replication of NDSS 2026 paper"
requires-python = ">=3.11"

dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "tenacity>=8.0.0",
    "youtube-search>=2.1.2",
    "rich>=13.0.0",       # pretty console output for logs
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chord"]